[MI-3254]: Added add config button in existing config table
