+++ webapp/src/containers/components/tables/styles.css
@@ -20,8 +20,9 @@ table.table thead {
     vertical-align: bottom;
 } 
 
-table {
+.table {
     width: 100%;
+
 }
 .table thead tr th td{
     white-space:nowrap
@@ -40,4 +41,8 @@ table {
 .btn-group .btn.btn-default:hover, .btn-group .btn.btn-default:focus {
     background: transparent;
     color: #000000;
+}
+
+.add-config-btn {
+    margin: 1em 0 0 0;
 }
