+++ webapp/src/index.tsx
@@ -8,7 +8,6 @@ import {PluginRegistry} from 'types/mattermostWebapp';
 import ExistingConfigTable from 'containers/components/tables/existingConfigTable';
 import ViewActionsModal from 'containers/components/modals/viewActionsModal';
 import DeleteConfigModal from 'containers/components/modals/deleteConfigModal';
-import AddConfigButton from 'containers/components/buttons/addConfigButton';
 
 import reducers from './reducers';
 
@@ -20,7 +19,6 @@ export default class Plugin {
         // @see https://developers.mattermost.com/extend/plugins/webapp/reference/
         registry.registerReducer(reducers);
         registry.registerAdminConsoleCustomSetting('ExistingConfigurationTable', ExistingConfigTable);
-        registry.registerAdminConsoleCustomSetting('AddConfigurationButton', AddConfigButton);
         registry.registerRootComponent(ViewActionsModal);
         registry.registerRootComponent(DeleteConfigModal);
     }
